server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "palaver"
version = "0.1.0"
#dynamic = ["version"]  # if you use version from git/tag, otherwise not needed

description = "A voice controlled toolset for LLM interaction"
#readme = "README.md"
requires-python = ">=3.12"
authors = [
    { name = "Dennis Parker", email = "dennis@windhavengroup.com" }
]


dependencies = [
    "aiomqtt>=2.0.0",
    "fastapi>=0.115.0",
    "jinja2>=3.1.6",
    "langchain>=1.2.0",
    "nicegui>=3.3.1",
    "numpy>=2.3.5",
    "ollama>=0.6.1",
    "packaging>=25.0",
    "pip>=25.3",
    "piper-tts>=1.3.0",
    "pyqt6>=6.10.1",
    "python-eventemitter>=1.0.13",
    "python-multipart>=0.0.20",
    "pyttsx3>=2.99",
    "pyyaml>=6.0.2",
    "qasync>=0.28.0",
    "rapidfuzz>=3.14.3",
    "requests>=2.32.5",
    "resampy>=0.4.3",
    "rich>=14.2.0",
    "scipy>=1.16.3",
    "silero-vad>=6.2.0",
    "sounddevice>=0.5.3",
    "soundfile>=0.13.1",
    "sqlmodel>=0.0.28",
    "textual>=6.7.1",
    "torch>=2.9.1",
    "torchaudio>=2.9.1",
    "tqdm>=4.67.1",
    "uvicorn>=0.32.0",
    "websockets>=14.0",
    "palaver-shared",
]
[tool.uv.sources]
palaver-shared= { path = "../packages/shared" , editable = true }
[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.4.2",
    "pytest-asyncio>=0.25.2",
    "pytest-cov>=7.0.0",
]

